==> 6.time_series_analysis_phu_dipwells.ipynb <==
%load_ext autoreload
%autoreload 2

import pandas as pd

# This dataframe contains all the field + explanatory variables (training data)
df = pd.read_csv("data/9_clean_training_data/all_training_data_with_extra_and_locations_and_precipSum.csv", parse_dates=["date"])
df

# Example list of numbers to filter the dataframe
phu_id_list = [371]  # Replace with your list of phu_id numbers

# Filtering the dataframe
filtered_df = df[df['phu_id'].isin(phu_id_list)]

import plotly.express as px

# Assuming df is already defined and contains the required data
# Convert 'date' column to datetime
df['date'] = pd.to_datetime(df['date'])

# Create the interactive time-series plot
fig = px.line(filtered_df, x='date', y='gwl_cm', color='id', title='Time Series of GWL by dipwell id',
              labels={'date': 'Date', 'gwl_cm': 'Ground Water Level (cm)'})

# Show the plot
fig.show()

# Example list of names to exclude from the 'id' variable
# exclude_id_list = ['BRG_140802_03']  # Replace with your list of ids to exclude
include_id_list = ['20_SPW_G23','20_SPW_D33','20_SPW_G29','20_SPW_B18']

# Further filter the dataframe by excluding specific ids
# filtered_df = filtered_df[~filtered_df['id'].isin(exclude_id_list)]
filtered_df = filtered_df[filtered_df['id'].isin(include_id_list)]

# Create the interactive time-series plot
fig = px.line(filtered_df, x='date', y='gwl_cm', color='id', title='Time Series of GWL by dipwell id',
              labels={'date': 'Date', 'gwl_cm': 'Ground Water Level (cm)'})

# Show the plot
fig.show()
